# contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neon Self-hosted Control API",
    "version": "0.5.0",
    "description": "Implemented self-hosted control API. M2 creation is enabled only when capability discovery reports its driver ready; unsupported SaaS Backend services remain disabled."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "neon_v2_session"
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ncp_ scoped API key",
        "description": "Expiry, revocation, current membership and key permission ceiling checked on every request."
      }
    },
...
    "schemas": {
...
      "Operation": {
        "type": "object",
        "required": [
          "id",
          "action",
          "state",
          "resource_id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "retry_wait",
              "succeeded",
              "failed"
            ]
          },
          "resource_type": {
            "type": "string"
          },
          "resource_id": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "error_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "retryable": {
            "type": "boolean"
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DataAPISpec": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "database",
          "schema",
          "issuer",
          "audience",
          "jwks",
          "allowed_origins"
        ],
        "properties": {
          "database": {
            "type": "string",
            "maxLength": 63
          },
          "schema": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{0,62}$"
          },
          "issuer": {
            "type": "string",
            "maxLength": 256
          },
          "audience": {
            "type": "string",
            "maxLength": 256
          },
          "jwks": {
            "type": "object",
            "description": "Static public RS256/EdDSA keys only; private parameters rejected"
          },
          "allowed_origins": {
            "type": "array",
            "maxItems": 32,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DataAPIInstance": {
        "type": "object",
        "required": [
          "branch_id",
          "generation",
          "state",
          "driver_enabled",
          "request_role",
          "public_endpoint"
        ],
        "properties": {
          "branch_id": {
            "type": "string"
          },
          "generation": {
            "type": "integer",
            "minimum": 0
          },
          "state": {
            "enum": [
              "disabled",
              "provisioning",
              "active",
              "disabling",
              "degraded"
            ]
          },
          "driver_enabled": {
            "type": "boolean"
          },
          "request_role": {
            "type": "string"
          },
          "public_endpoint": {
            "type": "string"
          },
          "spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataAPISpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "runtime": {
            "type": "object"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "endpoint_id": {
            "type": "string"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/projects/{project}/branches/{branch}/data-api": {
      "get": {
        "tags": [
          "Data API"
        ],
        "operationId": "getDataAPI",
        "summary": "Read configuration and runtime without SQL cold wake",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current instance; ETag is the generation",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataAPIInstance"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient project permission"
          },
          "404": {
            "description": "Project or branch not visible"
          },
          "409": {
            "description": "Conflicting active Operation, audience or service state"
          },
          "412": {
            "description": "Generation changed"
          },
          "422": {
            "description": "Invalid provider configuration"
          },
          "428": {
            "description": "If-Match required"
          },
          "503": {
            "description": "Driver disabled or metadata unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "Data API"
        ],
        "operationId": "enableDataAPI",
        "summary": "Enable branch Data API through an asynchronous native Driver",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Quoted numeric generation returned by GET; replay of the same key and input returns its original Operation"
          }
        ],
        "responses": {
          "202": {
            "description": "Stable durable Operation and branch ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "branch_id",
                    "operation"
                  ],
                  "properties": {
                    "branch_id": {
                      "type": "string"
                    },
                    "operation": {
                      "$ref": "#/components/schemas/Operation"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient project permission"
          },
          "404": {
            "description": "Project or branch not visible"
          },
          "409": {
            "description": "Conflicting active Operation, audience or service state"
          },
          "412": {
            "description": "Generation changed"
          },
          "422": {
            "description": "Invalid provider configuration"
          },
          "428": {
            "description": "If-Match required"
          },
          "503": {
            "description": "Driver disabled or metadata unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataAPISpec"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Data API"
        ],
        "operationId": "disableDataAPI",
        "summary": "Stop runtime and withdraw Proxy identity; retain application data",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Quoted numeric generation returned by GET; replay of the same key and input returns its original Operation"
          }
        ],
        "responses": {
          "202": {
            "description": "Stable durable Operation and branch ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "branch_id",
                    "operation"
                  ],
                  "properties": {
                    "branch_id": {
                      "type": "string"
                    },
                    "operation": {
                      "$ref": "#/components/schemas/Operation"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient project permission"
          },
          "404": {
            "description": "Project or branch not visible"
          },
          "409": {
            "description": "Conflicting active Operation, audience or service state"
          },
          "412": {
            "description": "Generation changed"
          },
          "422": {
            "description": "Invalid provider configuration"
          },
          "428": {
            "description": "If-Match required"
          },
          "503": {
            "description": "Driver disabled or metadata unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerToken": []
          }
        ]
      }
    },
...
  }
}